pyproject.toml
[project]
name = "whats-flying-over-my-house"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopy>=2.4.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
